requirements.txt
streamlit
transformers
torch
pytesseract
pdfplumber
python-docx
Pillow
keybert
sentence-transformers
langchain
scikit-learn
